refactor(shortener): remove unused analytics module and producer, move analytics service to services folder and inject MemphisService instead of using injection token test(shortener): add unit test for AnalyticsService test(shortener): update ShortenerController test to import MemphisModule and inject AnalyticsService feat(shortener): use AnalyticsService instead of AnalyticsProducer in ShortenerController to track requests and remove AnalyticsProducer injection from constructor
